It will ask for the keywords and the domain
This software will ask about the domain for which you are working and the keywords you want to check. 
This tool will save a lot of time of a SEO person who wants to consistently check the rankings of their client's website on certain keywords. 
This tool also has SQLite which will help track the keywords.